HotelBackend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
exponent_server_sdk==2.2.0
idna==3.11
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.2
python-decouple==3.8
